This commit is for Bugzilla 334836.  The Bugzilla contains patch 3 of 3 to add PPC64 LE support.  The other two patches can be found in Bugzillas 334384 and 334834.  Note, there are no VEX changes in this patch.
PP64 Little Endian test case fixes.

This patch adds new LE and BE expect files where needed.  In other
cases, the test was fixed to run correctly on LE and BE using based on
testing to see which platform is being used.

Where practical, the test cases have been changed so that the output
produced for BE	 and LE will be identical.  The test cases that require
a major rewrite to make the output identical for BE and LE simply
had an additional expect file added.

Signed-off-by: Carl Love <carll@us.ibm.com>