Power PC Add test suite support for ISA 3.0, part 1
The test suite support for the Power PC ISA 3.0 support is added in this commit.  The
following files were added:

   none/tests/ppc64/ppc64_helpers.h
   none/tests/ppc64/test_isa_3_0.c
   none/tests/ppc64/test_isa_3_0_altivec.stderr.exp
   none/tests/ppc64/test_isa_3_0_altivec.stdout.exp
   none/tests/ppc64/test_isa_3_0_altivec.vgtest
   none/tests/ppc64/test_isa_3_0_other.stderr.exp
   none/tests/ppc64/test_isa_3_0_other.stdout.exp
   none/tests/ppc64/test_isa_3_0_other.vgtest

The following file was modified:
   none/tests/ppc64/Makefile.am 

Note, the NEWS file was not updated as the ISA 3.0 support is not complete yet. This is the
first of five sets of commits needed to add all of the ISA 3.0 support.

valgrind bugzilla 359767

Corresponding Valgrind support added in:
  VEX commit 3214
  valgrind commit 15837